+++ scripts/04_probe_forward.py
@@ -41,10 +41,10 @@ inputs = tokenizer(text, return_tensors="pt").to(device)
 
 model = AutoModelForCausalLM.from_pretrained(
     model_id,
-    torch_dtype=dtype,
+    dtype=dtype,
     attn_implementation=cfg["attn_implementation"],
 )
-model.to(device)
+model.to(device)  # type: ignore[union-attr]
 model.eval()
 
 with torch.no_grad():
